Fix warnings about missing PNG support
Fix warnings about missing PNG support in Leptonica triggered by the
`DebugPixa` class. We aren't setting the debugging flags that cause this
class to actually produce useful output, but the warnings were still
being triggered when an instance of the class was constructed. Refactor
storage of `DebugPixa` so it isn't created at all when not needed.