Check for incorrect order of operations with OCREngine
Catch issues where an image is loaded before the training data/model or
OCR is attempted before an image is loaded. Tesseract doesn't report
errors at all or reliably for some of these operations, so use flags
within OCREngine to keep track of the state.

Also add tests for `getBoundingBoxes`.