Support recognition progress callbacks in OCRClient
Initially I tried to use the recommended approach to callbacks with comlink[1].
That ran into an issue with different instances of comlink being used in the
compiled dist/lib.js library and src/node-worker.js, which was fixed by moving
the `comlink.wrap` call into OCRClient. I then ran into an issue with the
proxied callback triggering a warning from Node in the worker thread
(`MaxListenersExceededWarning`). This happened because of the way that the
proxied callback on the worker side adds an event listener on each call, and
then only removes the listener after the callback response is received. If the
callback is invoked many times during the same event loop turn (or just before
the main thread has had a chance to respond) then the listeners accumulate and
in this case can easily hit Node's default warning limit of 10. To work around
that, `comlink.proxy(callback)` was replaced by creating a direct MessageChannel
ourselves for progress communication.

[1] https://github.com/GoogleChromeLabs/comlink#callbacks