Re-work progress delivery to work around Firefox issue
Create a single channel for all progress update notifications in the OCRClient
constructor and send the port to the worker ahead of any operations being
performed. This ensures that the port is fully ready by the time that
`OCREngine.{getText, getTextBoxes}` is invoked, and so any messages it sends are
dispatched immediately, rather than being deferred in the worker until the next
event loop turn, which is useless as by then all the work is done.