Pass image data to Tesseract as a Leptonica image
Reduce the number of copies of image data by creating a Leptonica image
and copying bytes from the ImageData object to the Leptonica buffer via
typed array APIs. This avoids an extra internal copy that Tesseract does
when a buffer pointer + length is passed to `TessBaseAPI::SetImage`. In
total I believe the image data is now copied twice:

 1. ImageData => Leptonica PIX* buffer
 2. TessBaseAPI::SetImage copies the input image

Tesseract operates on PIX* buffers internally, so potentially the second
copy could be avoided with changes to Tesseract.

Reducing copies is more efficient, and also reduces the peak memory usage,
allowing larger images to be processed given the same max memory allowance for
the WASM memory.