Free previous image before loading new one
Ask Tesseract to free its copy of the previous image before loading a new one.
This reduces the peak memory usage of the WASM module and reduces the
probability of hitting the current 128MB cap with large images.

When testing with a 2421x3307 scan of a document page in the web demo app for
example, recognition would succeed when the image was first loaded, but would
fail if you tried to load it a second time. Now the image can be reloaded
repeatedly.