Add simple orientation detection
Tesseract's has built-in script and orientation detection but it is part of the
classic (pre-LSTM) engine, which has been compiled out to reduce binary size.
Hence this initial implementation use Leptonica's more simplistic orientation
detection, which is based on counting numbers of ascenders and descenders, as
described on pages 12-14 of http://www.leptonica.org/papers/skew-measurement.pdf.