Display detected image orientation in OCR web demo
In adding this I encountered issues where the same rotated image dropped into
Safari, Chrome and Firefox could give different results. I believe this has to
do with how the EXIF rotation information is handled by the various browser APIs
used to load and draw images, but this still needs to be debugged.