Convert ImageBitmap => ImageData on the main thread in all browsers
Chrome has a bug where image orientation metadata in JPEG images is lost
when an ImageBitmap is cloned via a structured clone [1]. Therefore we have
to do ImageBitmap => ImageData conversion on the main thread to ensure
that the OCR engine receives decoded image data which respects the image
orientation.

Prior to this fix the rendered image orientation and the OCR output did
not match up in Chrome if the input image was rotated.

Since neither Firefox nor Safari support OffscreenCanvas, this means
that all browsers are now doing ImageBitmap => ImageData conversion on
the main thread.

Fixes #35

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1332947