Fix alignment of word box layer with canvas
The word box layer's height was slightly mismatched from the canvas where the
document image is drawn. This resulted in the word boxes drifting out of
alignment with the image towards the bottom. Fix this by setting a
`border-sizing` property to eliminate the mismatch in widths between the canvas
and box layer, and setting `aspect-ratio` on the box layer div to match the one
implicitly generated for the canvas by the browser.

Fixes #33